Split Platform resolve integration into tests that use java-platform plugin and those that do not
